pyproject.toml
[tool.poetry]
name = "catcident-backend"
version = "0.1.0"
description = ""
authors = ["2kwonhee <id@senterie.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
psycopg2-binary = "^2.9.10"
django-environ = "^0.11.2"
django-debug-toolbar = "^4.4.6"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.7.0"
django-ckeditor-5 = "^0.2.17"
django-parler = "^2.3"
django-storages = "^1.14.5"
boto3 = "^1.36.24"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
